WIP - update zuul packaging for zuul 2.0.0
This updates build dependencies so that a zuul 2.0.0 package can be
built.

At this time it is necessary to invoke sphinx without -W because
the two program-output directives in doc/source/client.rst cannot
succeed.

setup.cfg is patched to install the zuul-merger script, but since
zuul/cmd/merger.py is not part of zuul 2.0.0, it will not work.

Change-Id: I7b3cbf1fdfdad051633b4c32c86bfa5c8ab9b5df